make the transitory simulation not blocking. Results will be shown while sim analysis is still on going. Adapt the the speed of the animation to the user´s value if the animation takes too long.

+++ src/simulation/simulator.h
@@ -104,7 +104,7 @@ protected:
 	QHash<ConnectorItem *, int> m_connector2netHash;
 
 	QTimer *m_simTimer, *m_showResultsTimer;
-	int m_currSimStep;
+	unsigned long m_currSimStep;
 
 	static constexpr int SimDelay = 200;
 	static constexpr double HarmfulNegativeVoltage = -0.5;
